Enhance ksong action

+++ common/src/main/java/cn/booslink/llm/common/model/CBMSemantic.kt
@@ -46,7 +46,7 @@ data class Answer(val text: String?, val type: String?)
 class Semantic(
     @SerializedName("entrypoint") val entryPoint: String?,
     val intent: AIUIIntent?,
-    val score: Int?,
+    val score: Float?,
     val slots: List<Slot>,
     val template: String?
 )
